Use guide exercises for max tests
